Create Push Module and move stuff there
