+++ client/build.gradle
@@ -58,6 +58,7 @@ dependencies {
 
   api project(':core')
   api project(':core-android')
+  api project(':push')
   api project(':feature-users')
 
   implementation 'androidx.appcompat:appcompat:1.1.0-alpha04'
@@ -72,8 +73,6 @@ dependencies {
   implementation 'com.google.firebase:firebase-core:16.0.8'
   implementation 'com.google.firebase:firebase-config:16.4.1'
   implementation 'com.google.firebase:firebase-perf:16.2.4'
-  implementation 'com.google.firebase:firebase-messaging:17.5.0'
-  implementation 'com.google.firebase:firebase-database:16.1.0'
   implementation 'com.crashlytics.sdk.android:crashlytics:2.9.9'
 
   implementation 'androidx.lifecycle:lifecycle-runtime:2.0.0'
