+++ client/src/main/java/com/jraska/github/client/ui/RepoDetailActivity.kt
@@ -8,6 +8,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.airbnb.epoxy.SimpleEpoxyAdapter
 import com.airbnb.epoxy.SimpleEpoxyModel
 import com.jraska.github.client.R
+import com.jraska.github.client.core.android.BaseActivity
 import com.jraska.github.client.users.RepoDetail
 import com.jraska.github.client.users.RepoDetailViewModel
 import com.jraska.github.client.core.android.viewModel
