+++ core-android/src/main/AndroidManifest.xml
@@ -1,1 +1,19 @@
-<manifest package="com.jraska.github.client.core.android" />
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+  package="com.jraska.github.client.core.android">
+
+  <application>
+    <activity android:name=".UriHandlerActivity">
+      <intent-filter>
+        <action android:name="android.intent.action.VIEW"/>
+
+        <category android:name="android.intent.category.DEFAULT"/>
+        <category android:name="android.intent.category.BROWSABLE"/>
+
+        <data
+          android:host="github.com"
+          android:pathPattern="/.*"
+          android:scheme="https"/>
+      </intent-filter>
+    </activity>
+  </application>
+</manifest>
