+++ core-android/src/main/java/com/jraska/github/client/core/android/BaseActivity.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client.ui
+package com.jraska.github.client.core.android
 
 import androidx.appcompat.app.AppCompatActivity
