+++ core-android/src/main/java/com/jraska/github/client/core/android/UriHandlerActivity.kt
@@ -1,9 +1,6 @@
-package com.jraska.github.client.ui
+package com.jraska.github.client.core.android
 
 import android.os.Bundle
-import com.jraska.github.client.core.android.UriHandlerViewModel
-import com.jraska.github.client.inputUrl
-import com.jraska.github.client.core.android.viewModel
 
 class UriHandlerActivity : BaseActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
