+++ push/build.gradle
@@ -1,0 +1,28 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-kapt'
+
+android {
+  compileSdkVersion 28
+  defaultConfig {
+    minSdkVersion 21
+  }
+
+  compileOptions {
+    sourceCompatibility JavaVersion.VERSION_1_8
+    targetCompatibility JavaVersion.VERSION_1_8
+  }
+}
+
+dependencies {
+  api project(':core')
+  api project(':core-android')
+
+  api 'com.google.firebase:firebase-messaging:17.6.0'
+  api 'com.google.firebase:firebase-database:16.1.0'
+
+  implementation 'com.google.dagger:dagger:2.21'
+  kapt 'com.google.dagger:dagger-compiler:2.21'
+
+  testImplementation 'junit:junit:4.12'
+}
