+++ push/src/main/java/com/jraska/github/client/push/HasPushHandler.kt
@@ -1,0 +1,5 @@
+package com.jraska.github.client.push
+
+interface HasPushHandler {
+  fun pushHandler(): PushHandler
+}
