+++ push/src/main/java/com/jraska/github/client/push/SetAnalyticsPropertyPushCommand.kt
